revert

+++ torchscale/clip_utils/coco_segmentation.py
@@ -45,7 +45,7 @@ class COCOSegmentation:
             mask = self.target_transform(mask)
 
         # Convert mask to tensor and then to long type
-        mask = torch.from_numpy(np.array(mask))
+        mask = torch.from_numpy(np.array(mask)).long()
 
         return img, mask
